pyproject.toml
[project]
name = "fastapi-marketplace"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleg",email = "taranenko.oleg@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.135.1,<0.136.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "bandit[toml] (>=1.9.4,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [".venv", "alembic", "tests"]
skips = ["B101"]
